CCS-457: Modularize co2_containment.py
