+++ src/ccs_scripts/aggregate/_co2_mass.py
@@ -9,8 +9,8 @@ from ccs_scripts.aggregate._config import CO2MassSettings
 from ccs_scripts.co2_containment.co2_calculation import (
     Co2Data,
     Co2DataAtTimeStep,
-    Scenario,
 )
+from ccs_scripts.co2_containment.source_data import Scenario
 from ccs_scripts.utils.timer import Timer
 from ccs_scripts.utils.xtgeo_logging import setup_xtgeo_logging
